Refactor to reuse getAdminSetting code
